requirements.txt
openai
psycopg2-binary
pgvector
pydantic
python-docx
python-dotenv